refactor(tx-flow): move ProposerForm component to tx-flow folder
